style: separate BlockContents module
