+++ notion_df/request/page.py
@@ -6,7 +6,8 @@ from uuid import UUID
 
 from notion_df.core.request import SingleRequestBuilder, RequestSettings, Version, Method, PaginatedRequestBuilder, \
     RequestBuilder
-from notion_df.object.data import BlockContents, serialize_block_value_list, PageData
+from notion_df.object.contents import BlockContents
+from notion_df.object.data import serialize_block_value_list, PageData
 from notion_df.object.file import ExternalFile
 from notion_df.object.misc import Icon, PartialParent
 from notion_df.property import PageProperties, Property, property_registry, PagePropertyValueT
